Implement all utils; test import rewriting

+++ src/main.ts
@@ -1,43 +1,50 @@
-import { glob } from 'npm:glob';
 import { out } from './out.ts';
 import {
+  getPaths,
+  isEligibleDir,
+  isLoneFile,
   logError,
-  moveFileUp,
   removeDir,
-  rewriteRelativeImports,
-  verifyIsEligibleDir,
-  verifyIsLoneFile,
+  rewriteFile,
 } from './utils.ts';
 
 const TIME_LOG_KEY = 'Processing time';
+const DEBUG = Deno.args.includes('--debug');
 
 export default async function main() {
   console.time(TIME_LOG_KEY);
 
-  const paths = await glob('src/**/index.{ts,tsx,js,jsx}');
-
-  const files = new Set();
+  const paths = await getPaths();
+  let pathsCount = 0;
+  const successPaths = new Set();
+  const skippedPaths = new Set();
   const errors = new Map();
 
-  for (const path of paths) {
+  for await (const { path, isFile } of paths) {
+    pathsCount++;
+    out.log(`- ${path}`);
+
     try {
-      verifyIsEligibleDir(path);
-      await verifyIsLoneFile(path);
-      await moveFileUp(path);
-      await rewriteRelativeImports(path);
+      if (!isFile || !isEligibleDir(path) || !(await isLoneFile(path))) {
+        skippedPaths.add(path);
+        continue;
+      }
+
+      await rewriteFile(path);
       await removeDir(path);
-      files.add(path);
+
+      successPaths.add(path);
     } catch (error) {
       logError(path, error, errors);
     }
+    break; // TODO Debug remove
   }
 
   out.log('—'.repeat(40));
   console.timeEnd(TIME_LOG_KEY);
-
-  out.log(`Total files found: ${paths.length}`);
-  out.success(`Total files moved: ${files.size}`);
-
+  out.log(`Total files found: ${pathsCount}`);
+  out.log(`Skipped files: ${skippedPaths.size}`);
+  out.success(`Total files moved: ${successPaths.size}`);
   if (errors.size > 0) {
     out.error(`Encountered errors: ${errors.size}`);
   }
